MINI_AI/requirements-gpu.txt
# Caelonyx AI Stack - GPU-Enabled Dependencies
torch
torchvision
cupy-cuda12x
faiss-gpu
numpy
gradio
